// components/Shared/Typography/types.ts
export type Variant =
  | "headline-bold"
  | "headline-light"
  | "subhead-medium"
  | "subhead-medium-small"
  | "subhead-regular"
  | "body-bold"
  | "body-regular"
  | "body-medium"
  | "caption-regular"
  | "caption-light"
  | "caption-medium"
  | "digit"
  | "english-regular"
  | "english-regular-small"
  | "button";